Multi-Route Selection for Carrier; Basic Time Simulation
CarrierList functions now return nested Lists of Trips so that the Carrier can select an ideal route instead of having TMS select the first option on their behalf. Updated all (?) comments regarding this to reflect the change.

Created a basic time simulation function in the order -- really doesn't belong there but it's the easiest place to implement it so c'est la vie.